requirements.txt
numpy==1.19.2
Pillow==7.2.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
PyYAML~=5.3.1
tqdm~=4.50.0